# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog API",
    "description": "API REST complète pour la gestion d'un blog avec authentification JWT et CRUD d'articles",
    "version": "1.0.0",
    "contact": {
      "name": "Votre Nom",
      "email": "votre.email@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Serveur de développement"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Article": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID unique de l'article"
          },
          "title": {
            "type": "string",
            "description": "Titre de l'article"
          },
          "slug": {
            "type": "string",
            "description": "Slug unique de l'article"
          },
          "content": {
            "type": "string",
            "description": "Contenu de l'article"
          },
          "excerpt": {
            "type": "string",
            "description": "Extrait de l'article"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags de l'article"
          },
          "category": {
            "type": "string",
            "enum": [
              "technologie",
              "lifestyle",
              "voyage",
              "cuisine",
              "sport",
              "autre"
            ],
            "description": "Catégorie de l'article"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "published", "archived"],
            "default": "draft",
            "description": "Statut de l'article"
          },
          "featuredImage": {
            "type": "string",
            "description": "URL de l'image principale"
          },
          "readTime": {
            "type": "integer",
            "description": "Temps de lecture en minutes"
          },
          "views": {
            "type": "integer",
            "default": 0,
            "description": "Nombre de vues"
          },
          "likes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "description": "Utilisateurs qui ont liké l'article"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "user": {
                  "$ref": "#/components/schemas/User"
                },
                "content": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "seo": {
            "type": "object",
            "properties": {
              "metaTitle": {
                "type": "string"
              },
              "metaDescription": {
                "type": "string"
              },
              "keywords": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/articles": {
      "get": {
        "summary": "Récupérer les articles",
        "description": "Récupérer les articles de l'utilisateur connecté avec pagination et filtres",
        "tags": ["Articles"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Numéro de page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Nombre d'éléments par page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filtrer par catégorie",
            "schema": {
              "type": "string",
              "enum": [
                "technologie",
                "lifestyle",
                "voyage",
                "cuisine",
                "sport",
                "autre"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrer par statut",
            "schema": {
              "type": "string",
              "enum": ["draft", "published", "archived"]
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Recherche textuelle",
            "schema": {
              "type": "string",
              "minLength": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Articles récupérés avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "articles": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Article"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "currentPage": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            },
                            "totalItems": {
                              "type": "integer"
                            },
                            "itemsPerPage": {
                              "type": "integer"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "hasPrevPage": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Créer un article",
        "description": "Créer un nouvel article",
        "tags": ["Articles"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content", "category"],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 200
                  },
                  "content": {
                    "type": "string",
                    "minLength": 50
                  },
                  "excerpt": {
                    "type": "string",
                    "maxLength": 300
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "technologie",
                      "lifestyle",
                      "voyage",
                      "cuisine",
                      "sport",
                      "autre"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "published", "archived"],
                    "default": "draft"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 20
                    }
                  },
                  "featuredImage": {
                    "type": "string",
                    "format": "uri"
                  },
                  "seo": {
                    "type": "object",
                    "properties": {
                      "metaTitle": {
                        "type": "string",
                        "maxLength": 60
                      },
                      "metaDescription": {
                        "type": "string",
                        "maxLength": 160
                      },
                      "keywords": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Article créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "article": {
                          "$ref": "#/components/schemas/Article"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erreur de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/articles/public": {
      "get": {
        "summary": "Récupérer les articles publics",
        "description": "Récupérer les articles publiés (accès public)",
        "tags": ["Articles"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Numéro de page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Nombre d'éléments par page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20,
              "default": 10
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filtrer par catégorie",
            "schema": {
              "type": "string",
              "enum": [
                "technologie",
                "lifestyle",
                "voyage",
                "cuisine",
                "sport",
                "autre"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Recherche textuelle",
            "schema": {
              "type": "string",
              "minLength": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Articles publics récupérés avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "articles": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Article"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "currentPage": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            },
                            "totalItems": {
                              "type": "integer"
                            },
                            "itemsPerPage": {
                              "type": "integer"
                            },
                            "hasNextPage": {
                              "type": "boolean"
                            },
                            "hasPrevPage": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/articles/{id}": {
      "get": {
        "summary": "Récupérer un article",
        "description": "Récupérer un article spécifique par ID",
        "tags": ["Articles"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de l'article",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Article récupéré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "article": {
                          "$ref": "#/components/schemas/Article"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Article non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Mettre à jour un article",
        "description": "Mettre à jour un article existant",
        "tags": ["Articles"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de l'article",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 200
                  },
                  "content": {
                    "type": "string",
                    "minLength": 50
                  },
                  "excerpt": {
                    "type": "string",
                    "maxLength": 300
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "technologie",
                      "lifestyle",
                      "voyage",
                      "cuisine",
                      "sport",
                      "autre"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "published", "archived"]
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 20
                    }
                  },
                  "featuredImage": {
                    "type": "string",
                    "format": "uri"
                  },
                  "seo": {
                    "type": "object",
                    "properties": {
                      "metaTitle": {
                        "type": "string",
                        "maxLength": 60
                      },
                      "metaDescription": {
                        "type": "string",
                        "maxLength": 160
                      },
                      "keywords": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article mis à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "article": {
                          "$ref": "#/components/schemas/Article"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Non autorisé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Article non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Supprimer un article",
        "description": "Supprimer un article",
        "tags": ["Articles"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID de l'article",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Article supprimé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Non autorisé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Article non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}